Create Boutique Store Database
My very first database/project through Khan Academy where I designed a "boutique" and imagined what items would be sold and how much as well as having a rating to the items. Looking back, I would have broken it down to the customer ID and which items were bought and individual ratings to them to further be able to push "findings" with the data further to make it feel a bit more realistic. However, it was definitely rewarding and challenging at the same time.